added list product functionality and deprecated old product form
