+++ pages/marketplace/[[...pubkey]].tsx
@@ -4,7 +4,6 @@ import { useRouter } from "next/router";
 import { ArrowUturnLeftIcon } from "@heroicons/react/24/outline";
 import { nip19, SimplePool } from "nostr-tools";
 import ProductForm from "../components/product-form";
-import NewForm from "../components/NewForm";
 
 const SellerView = () => {
   const router = useRouter();
@@ -80,7 +79,10 @@ const SellerView = () => {
         showModal={showModal}
         handleModalToggle={handleModalToggle}
       /> */}
-      <NewForm showModal={showModal} handleModalToggle={handleModalToggle} />
+      <ProductForm
+        showModal={showModal}
+        handleModalToggle={handleModalToggle}
+      />
     </div>
   );
 };
